Create Get The Gift!
Today is the dog’s (aka your) birthday! It was a great party and the first gift you opened was a propeller hat! Suddenly…the gifts grew wings and flew away?! Turns out the evil cat (player two) set out to get you! In this game you have to maneuver through clouds and avoid the evil cat to get your gifts! This game is a two player game and can provide fun (and arguments…) enjoy!